py-src/requirements.txt
Babel==2.12.1
emoji-country-flag==1.3.2
